// src/app/models/user.ts
// export interface ProfileUser {
//     uid: string;
//     nombre?: string;
//     apellido?: string;
//     displayName?: string;
//     telefono?: string;
//     rol?: string;
//     email?: string;
// }

export interface ProfileUser {
    uid: string;
    email?: string;
    firstName?: string;
    lastName?: string;
    displayName?: string;
    phone?: string;
    address?: string;
    photoURL?: string;
    rol?: string;
  }